pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "passgen-core"
version = "0.1.0"
description = "Core utilities for PassGen — password generation and secure primitives."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "argon2-cffi>=21.3.0",
  "pycryptodome>=3.18.0"
]
authors = [{name = "Echo"}]
license = {text = "MIT"}